+++ fabric-model-events-api-v1/src/client/java/net/fabricmc/fabric/mixin/client/modelevents/EntityRenderDispatcherMixin.java
@@ -1,0 +1,29 @@
+package net.fabricmc.fabric.mixin.client.modelevents;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import net.fabricmc.fabric.impl.client.modelevents.ModelRenderContext;
+import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.render.entity.EntityRenderDispatcher;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.Entity;
+
+@Mixin(EntityRenderDispatcher.class)
+abstract class EntityRenderDispatcherMixin {
+    private static final String RENDER = "render(Lnet/minecraft/entity/Entity;DDDFFLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;I)V";
+
+    @Inject(method = RENDER, at = @At("HEAD"))
+    private <E extends Entity> void before_render(E entity, double x, double y, double z, float yaw, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, CallbackInfo info) {
+        ModelRenderContext.CURRENT_ENTITY.push(entity);
+    }
+
+    @Inject(method = RENDER, at = @At("RETURN"))
+    private <E extends Entity> void after_render(E entity, double x, double y, double z, float yaw, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, CallbackInfo info) {
+        if (!ModelRenderContext.CURRENT_ENTITY.isEmpty()) {
+            ModelRenderContext.CURRENT_ENTITY.pop();
+        }
+    }
+}
