+++ fabric-model-events-api-v1/src/client/resources/assets/fabric-mining-level-api-v1/icon.png
