+++ fabric-model-events-api-v1/src/client/resources/fabric.mod.json
