--- fabric-model-events-api-v1/src/main/resources/fabric-model-events-api-v1.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.modelevents",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "BlockEntityRenderDispatcherMixin",
-    "EntityRenderDispatcherMixin",
-    "ModelPart_CuboidMixin",
-    "ModelPart_QuadMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
